[sui-fork] Centralize fork operations in Context
Move clock, checkpoint, and status behavior out of the gRPC transport so future in-process callers can share the same implementation without introducing another control type.
